requirement.txt
asgiref==3.2.7
asn1crypto==0.24.0
attrs==19.3.0
Automat==0.6.0
blinker==1.4
certifi==2020.4.5.2
chardet==3.0.4
click==7.1.2
cloud-init==19.4
colorama==0.3.7
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.1.4
dateutils==0.6.8
distro-info===0.18ubuntu0.18.04.1
Django==3.0.7
django-braces==1.13.0
django-cors-headers==3.3.0
django-cors-middleware==1.5.0
django-discover-runner==1.0
django-filter==2.2.0
django-nose==1.4.6
django-oauth-toolkit==1.3.2
django-role-permissions==2.2.1
django-rq==2.2.0
djangorestframework==3.11.0
httplib2==0.9.2
hyperlink==17.3.1
idna==2.9
importlib-metadata==1.6.1
incremental==16.10.1
jdatetime==3.4.0
jedi==0.17.0
Jinja2==2.11.2
joblib==0.14.1
json5==0.9.4
jsonpatch==1.16
jsonpointer==1.10
jsonschema==3.2.0
kavenegar==1.1.2
Khayyam==3.0.17
language-selector==0.1
MarkupSafe==1.1.1
mongoengine==0.20.0
mongoengine-goodjson==1.1.8
netifaces==0.10.4
nose==1.3.7
oauthlib==3.1.0
PAM==0.4.2
parso==0.7.0
passlib==1.7.2
psycopg2-binary==2.8.5
pyasn1==0.4.2
pyasn1-modules==0.2.1
pygobject==3.26.1
PyJWT==1.5.3
pymongo==3.10.1
pyOpenSSL==17.5.0
pyrsistent==0.16.0
pyserial==3.4
python-apt==1.6.5+ubuntu0.2
python-dateutil==2.8.1
python-debian==0.1.32
pytz==2020.1
PyYAML==3.12
redis==3.5.3
requests==2.23.0
requests-unixsocket==0.1.5
rq==1.4.2
service-identity==16.0.0
six==1.15.0
sqlparse==0.3.1
ssh-import-id==5.7
systemd-python==234
Twisted==17.9.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.9
zipp==3.1.0
zope.interface==4.3.2